requirements.txt
requests
numpy
python-dateutil